fix: state값 초기화 오류
