+++ src/apis/auth.ts
@@ -11,7 +11,7 @@ interface GenericResponse {
 export const checkLogin = (social: string, token: string) =>
   Api.post(`/oauth/login/${social}`, token);
 
-export const userLogin = (data: any, token: string) => {
+export const signUp = (data: any, token: string) => {
   return Api.post(`/oauth/signup`, data, {
     headers: {
       signUpToken: token,
